+++ packages/shared/src/provisioning/index.ts
@@ -1,3 +1,4 @@
 export * from './constants';
 export * from './validate';
 export * from './plan';
+export * from './deprovision';